cody: add comments to completions modules (#53205)
- No functional changes. I only added comments, renamed a couple of
properties and moved files around.
- Follow-up draft https://github.com/sourcegraph/sourcegraph/pull/53223
