+++ client/cody/package.json
@@ -8,7 +8,7 @@
   "icon": "resources/cody.png",
   "description": "AI code assistant that writes code and answers questions for you",
   "scripts": {
-    "start:debug": "NODE_ENV=development CODY_FOCUS_ON_STARTUP=1 code --extensionDevelopmentPath=\"$INIT_CWD/client/cody\" --disable-extension=sourcegraph.cody-ai --inspect-extensions=9333 . --goto ./src/local-app-detector.ts:12:59",
+    "start:debug": "NODE_ENV=development code --extensionDevelopmentPath=\"$INIT_CWD/client/cody\" --disable-extension=sourcegraph.cody-ai --inspect-extensions=9333 . --goto ./src/local-app-detector.ts:12:59",
     "build": "pnpm esbuild --minify && vite build --mode production",
     "build:dev": "concurrently \"pnpm esbuild --sourcemap\" \"vite build --mode development\"",
     "download-rg": "scripts/download-rg.sh",
