+++ client/cody/src/completions/prompts.ts
@@ -134,9 +134,9 @@ export class KnowledgeBasePromptTemplate implements PromptTemplate {
                 {
                     speaker: 'human',
                     text:
-                        `Add the following code snippet (from file ${snippet.filename}) to your knowledge base:\n` +
+                        `Add the following code snippet (from file ${snippet.fileName}) to your knowledge base:\n` +
                         '```' +
-                        `\n${snippet.text}\n` +
+                        `\n${snippet.content}\n` +
                         '```',
                 },
                 {
